Change timeZoneName to 'long' to show location and not just offset

+++ react/components/station_formats.js
@@ -92,7 +92,7 @@ export function formatCioosDateTime(date_str){
     day: 'numeric', 
     hour: '2-digit', 
     minute: '2-digit',
-    timeZoneName: 'short'
+    timeZoneName: 'long'
 };
 
   const timestamp = date.toLocaleString('en-US', options);
